Revert "Add implementation of the API methods"
This reverts commit 6ba9d94b7a73935119585c832c921ebe34882cfc.
